Update README.md
This is a collaboration project with dirkessl (and likely others) on a new and improved version of the LED Image Painter project.
This is a similar device to the commercial PixelStick. It displays bmp files and built-in patterns in the air to be captured with time exposures.
The images are displayed using one or two 144 LED RGB pixel arrays.
The processor is currently the TTGO with the TFT display. T-Display.
It has a nice menu system and single button for control.